server/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
WTForms==2.3.3
python-dotenv==0.19.2
SQLAlchemy==1.4.25
Werkzeug==2.2.2
pymysql==1.0.2